Add auto switch uplink
