+++ .gitignore
@@ -7,6 +7,7 @@
 *.pyc
 .idea/encodings.xml
 *.png
+*.jpg
 *.flv
 *.mp4
 *.part
