+++ engine/plugins/upload/bili_web.py
@@ -1,6 +1,7 @@
 import base64
 import hashlib
 import json
+import math
 import os
 import time
 from dataclasses import dataclass, field, InitVar, asdict
@@ -11,19 +12,14 @@ from urllib.parse import quote
 
 import requests
 import rsa
-import selenium.common
+
 from requests import utils
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.support import expected_conditions as ec
-from selenium.webdriver.support.ui import WebDriverWait
+
 
 import engine
 from common.decorators import Plugin
-from engine.plugins import logger
-from engine.plugins.base_adapter import UploadBase
-from engine.slider import slider_cracker
+
+from engine.plugins.upload import UploadBase, logger
 
 
 @Plugin.upload(platform="bili_web")
@@ -55,7 +51,7 @@ class BiliWeb(UploadBase):
             if img_path:
                 video.cover = bili.cover_up(img_path).replace('http:', '')
             ret = bili.submit()  # 提交视频
-        logger.info(f"upload_success:{ret}")
+        logger.info(f"上传成功: {ret}")
         self.remove_filelist(file_list)
 
     def login(self, b):
@@ -79,196 +75,6 @@ class BiliWeb(UploadBase):
             json.dump(cookies, f)
 
 
-@Plugin.upload("bilibili")
-class BiliChrome(UploadBase):
-    def __init__(self, principal, data):
-        super().__init__(principal, data, 'engine/bilibili.cookie')
-        # self.title = title
-        # self.date_title = None
-        self.driver = None
-
-    @staticmethod
-    def assemble_videopath(file_list):
-        root = os.getcwd()
-        videopath = ''
-        for i in range(len(file_list)):
-            file = file_list[i]
-            videopath += root + '/' + file + '\n'
-        videopath = videopath.rstrip()
-        return videopath
-
-    @staticmethod
-    def is_element_exist(driver, xpath):
-        s = driver.find_elements_by_xpath(xpath=xpath)
-        if len(s) == 0:
-            print("元素未找到:%s" % xpath)
-            return False
-        elif len(s) == 1:
-            return True
-        else:
-            print("找到%s个元素：%s" % (len(s), xpath))
-            return False
-
-    def upload(self, file_list):
-
-        filename = self.persistence_path
-        videopath = self.assemble_videopath(file_list)
-
-        # service_log_path = "{}/chromedriver.log".format('/home')
-        options = webdriver.ChromeOptions()
-
-        options.add_argument('headless')
-        self.driver = webdriver.Chrome(executable_path=engine.chromedriver_path, chrome_options=options)
-        # service_log_path=service_log_path)
-        try:
-            self.driver.get("https://www.bilibili.com")
-            # driver.delete_all_cookies()
-            if os.path.isfile(filename):
-                with open(filename) as f:
-                    new_cookie = json.load(f)
-
-                for cookie in new_cookie:
-                    if isinstance(cookie.get("expiry"), float):
-                        cookie["expiry"] = int(cookie["expiry"])
-                    self.driver.add_cookie(cookie)
-
-            self.driver.get("https://member.bilibili.com/video/upload.html")
-
-            # print(driver.title)
-            self.add_videos(videopath)
-
-            # js = "var q=document.getElementsByClassName('content-header-right')[0].scrollIntoView();"
-            # driver.execute_script(js)
-
-            cookie = self.driver.get_cookies()
-            with open(filename, "w") as f:
-                json.dump(cookie, f)
-
-            self.add_information()
-
-            self.driver.find_element_by_xpath('//*[@class="upload-v2-container"]/div[2]/div[3]/div[5]/span[1]').click()
-            # screen_shot = driver.save_screenshot('bin/1.png')
-            # print('截图')
-            time.sleep(3)
-            upload_success = self.driver.find_element_by_xpath(r'//*[@id="app"]/div/div[3]/h3').text
-            if upload_success == '':
-                self.driver.save_screenshot('err.png')
-                logger.info('稿件提交失败，截图记录')
-                return
-            else:
-                logger.info(upload_success)
-            # logger.info('%s提交完成！' % title_)
-            self.remove_filelist(file_list)
-        except selenium.common.exceptions.NoSuchElementException:
-            logger.exception('发生错误')
-        # except selenium.common.exceptions.TimeoutException:
-        #     logger.exception('超时')
-        except selenium.common.exceptions.TimeoutException:
-            self.login(filename)
-
-        finally:
-            self.driver.quit()
-            logger.info('浏览器驱动退出')
-
-    def login(self, filename):
-        logger.info('准备更新cookie')
-        # screen_shot = driver.save_screenshot('bin/1.png')
-        WebDriverWait(self.driver, 10).until(
-            ec.presence_of_element_located((By.XPATH, r'//*[@id="login-username"]')))
-        username = self.driver.find_element_by_xpath(r'//*[@id="login-username"]')
-        username.send_keys(engine.config['user']['account']['username'])
-        password = self.driver.find_element_by_xpath('//*[@id="login-passwd"]')
-        password.send_keys(engine.config['user']['account']['password'])
-        self.driver.find_element_by_class_name("btn-login").click()
-        # logger.info('第四步')
-        # try:
-        cracker = slider_cracker(self.driver)
-        cracker.crack()
-        # except:
-        #     logger.exception('出错')
-        time.sleep(5)
-        if self.driver.title == '投稿 - 哔哩哔哩弹幕视频网 - ( ゜- ゜)つロ 乾杯~ - bilibili':
-            cookie = self.driver.get_cookies()
-            print(cookie)
-            with open(filename, "w") as f:
-                json.dump(cookie, f)
-            logger.info('更新cookie成功')
-        else:
-            logger.info('更新cookie失败')
-
-    def add_videos(self, videopath):
-        formate_title = self.data["format_title"]
-        WebDriverWait(self.driver, 20).until(
-            ec.presence_of_element_located((By.NAME, 'buploader')))
-        upload = self.driver.find_element_by_name('buploader')
-        # logger.info(driver.title)
-        upload.send_keys(videopath)  # send_keys
-        logger.info('开始上传' + formate_title)
-        time.sleep(2)
-        button = r'//*[@class="new-feature-guide-v2-container"]/div/div/div/div/div[1]'
-        if self.is_element_exist(self.driver, button):
-            sb = self.driver.find_element_by_xpath(button)
-            sb.click()
-            sb.click()
-            sb.click()
-            logger.debug('点击')
-        while True:
-            try:
-                info = self.driver.find_elements_by_class_name(r'item-upload-info')
-                for t in info:
-                    if t.text != '':
-                        print(t.text)
-                time.sleep(10)
-                text = self.driver.find_elements_by_xpath(r'//*[@class="item-upload-info"]/span')
-                aggregate = set()
-                for s in text:
-                    if s.text != '':
-                        aggregate.add(s.text)
-                        print(s.text)
-
-                if len(aggregate) == 1 and ('Upload complete' in aggregate or '上传完成' in aggregate):
-                    break
-            except selenium.common.exceptions.StaleElementReferenceException:
-                logger.exception("selenium.common.exceptions.StaleElementReferenceException")
-        logger.info('上传%s个数%s' % (formate_title, len(info)))
-
-    def add_information(self):
-        link = self.data.get("url")
-        # 点击模板
-        self.driver.find_element_by_xpath(r'//*[@class="normal-title-wrp"]/div/p').click()
-        self.driver.find_element_by_class_name(r'template-list-small-item').click()
-        # driver.find_element_by_xpath(
-        #     r'//*[@id="app"]/div[3]/div[2]/div[3]/div[1]/div[1]/div/div[2]/div[1]').click()
-        # 输入转载来源
-        input_o = self.driver.find_element_by_xpath(
-            '//*[@class="upload-v2-container"]/div[2]/div[3]/div[1]/div[4]/div[3]/div/div/input')
-        input_o.send_keys(link)
-        # 选择分区
-        # driver.find_element_by_xpath(r'//*[@id="item"]/div/div[2]/div[3]/div[2]/div[2]/div[1]/div[2]/div[2]/div[1]/div[3]/div').click()
-        # driver.find_element_by_xpath(r'//*[@id="item"]/div/div[2]/div[3]/div[2]/div[2]/div[1]/div[2]/div[2]/div[1]/div[3]/div[2]/div[6]').click()
-        # 稿件标题
-        title = self.driver.find_element_by_xpath(
-            '//*[@class="upload-v2-container"]/div[2]/div[3]/div[1]/div[8]/div[2]/div/div/input')
-        title.send_keys(Keys.CONTROL + 'a')
-        title.send_keys(Keys.BACKSPACE)
-        title.send_keys(self.data["format_title"])
-        # js = "var q=document.getElementsByClassName('content-tag-list')[0].scrollIntoView();"
-        # driver.execute_script(js)
-        # time.sleep(3)
-        # 输入相关游戏
-        # driver.save_screenshot('bin/err.png')
-        # print('截图')
-        # text_1 = driver.find_element_by_xpath(
-        #     '//*[@id="item"]/div/div[2]/div[3]/div[2]/div[2]/div[1]/div[5]/div/div/div[1]/div[2]/div/div/input')
-        # text_1.send_keys('星际争霸2')
-        # 简介
-        text_2 = self.driver.find_element_by_xpath(
-            '//*[@class="upload-v2-container"]/div[2]/div[3]/div[1]/div[12]/div[2]/div/textarea')
-        text_2.send_keys('职业选手直播第一视角录像。这个自动录制上传的小程序开源在Github：'
-                         'http://t.cn/RgapTpf(或者在Github搜索ForgQi)\n'
-                         '交流群：837362626')
-
-
 class BiliBili:
     def __init__(self, video: 'Data'):
         self.app_key = 'bca7e84c2d947ac6'
@@ -280,6 +86,7 @@ class BiliBili:
             "Referer": "https://www.bilibili.com/", 'Connection': 'keep-alive'
         })
         self.__bili_jct = None
+        self._auto_os = None
 
     def login_by_password(self, username, password):
         print('使用账号上传')
@@ -340,18 +147,100 @@ class BiliBili:
         if r and r["code"] == 0:
             return r['data']['hash'], rsa.PublicKey.load_pkcs1_openssl_pem(r['data']['key'].encode())
 
-    def upload_file(self, filepath: str):
-        """上传本地视频文件,返回视频信息dict, fsize=8388608"""
-        import math
-        path, name = os.path.split(filepath)  # 分离路径与文件名
+    def probe(self):
+        ret = self.__session.get('https://member.bilibili.com/preupload?r=probe', timeout=5).json()
+        data, auto_os = None, None
+        min_cost = 0
+        if ret['probe'].get('get'):
+            method = 'get'
+        else:
+            method = 'post'
+            data = bytes(int(1024 * 0.1 * 1024))
+        for line in ret['lines']:
+            start = time.perf_counter()
+            test = self.__session.request(method, f"https:{line['probe_url']}", data=data, timeout=30)
+            cost = time.perf_counter() - start
+            print(line['query'], cost)
+            if test.status_code != 200:
+                return
+            if not min_cost or min_cost > cost:
+                auto_os = line
+                min_cost = cost
+        return auto_os
 
+    def upload_file(self, filepath: str):
+        """上传本地视频文件,返回视频信息dict
+        b站目前支持4种上传线路upos, kodo, gcs, bos
+        kodo: {"os":"kodo","query": "bucket=bvcupcdnkodobm&probe_version=20200810",
+        "probe_url":"//up-na0.qbox.me/crossdomain.xml"}
+        gcs: {"os":"gcs","query":"bucket=bvcupcdngcsus&probe_version=20200810",
+        "probe_url":"//storage.googleapis.com/bvcupcdngcsus/OK"},
+        upos:
+        {"os":"upos","query":"upcdn=ws&probe_version=20200810","probe_url":"//upos-sz-upcdnws.bilivideo.com/OK"}
+        {"os":"upos","query":"upcdn=qn&probe_version=20200810","probe_url":"//upos-sz-upcdnqn.bilivideo.com/OK"}
+        {"os":"upos","query":"upcdn=bda2&probe_version=20200810","probe_url":"//upos-sz-upcdnbda2.bilivideo.com/OK"}
+        """
+        if not self._auto_os:
+            self._auto_os = self.probe()
+            # self._auto_os = {"os": "kodo", "query": "bucket=bvcupcdnkodobm&probe_version=20200810",
+            #                  "probe_url": "//up-na0.qbox.me/crossdomain.xml"}
+            logger.info(f"开始上传{self._auto_os['os']}: {self._auto_os['query']}")
+        profile = 'ugcupos/bup' if 'upos' == self._auto_os['os'] else "ugcupos/bupfetch"
+        query = f"r={self._auto_os['os']}&profile={quote(profile, safe='')}" \
+                f"&ssl=0&version=2.8.9&build=2080900&{self._auto_os['query']}"
+        if self._auto_os['os'] == 'upos':
+            return self.upos(filepath, query)
+        elif self._auto_os['os'] == 'kodo':
+            return self.kodo(filepath, query)
+        elif self._auto_os['os'] == "gcs":
+            raise NotImplementedError('gcs')
+        elif self._auto_os['os'] == "bos":
+            raise NotImplementedError('bos')
+        else:
+            logger.error(f"NoSearch:{self._auto_os['os']}")
+
+    def kodo(self, filepath, query):
+        total = os.path.getsize(filepath)
+        chunk_size = 4194304
         with open(filepath, 'rb') as f:
-            # size = f.seek(0, 2)  # 获取文件大小
-            total = os.path.getsize(filepath)
-            # 申请上传返回上传信息
-            ret = self.__session.get(f'https://member.bilibili.com/preupload?name={quote(name)}&size={total}'
-                                     f'&r=upos&profile=ugcupos%2Fbup&ssl=0&version=2.8.9'
-                                     f'&build=2080900&upcdn=bda2&probe_version=20200628', timeout=5).json()
+            name = f.name
+            ret = self.__session.get(
+                f'https://member.bilibili.com/preupload?name={quote(name)}&size={total}&{query}', timeout=5).json()
+            bili_filename = ret['bili_filename']
+            key = ret['key']
+            endpoint = f"https:{ret['endpoint']}"
+            token = ret['uptoken']
+            fetch_url = ret['fetch_url']
+            fetch_headers = ret['fetch_headers']
+            url = f'{endpoint}/mkblk'
+            headers = {
+                'Authorization': f"UpToken {token}",
+            }
+            # 开始上传
+            parts = []  # 分块信息
+            chunks = math.ceil(total / chunk_size)  # 获取分块数量
+            for i in range(chunks):
+                chunks_data = f.read(chunk_size)  # 一次读取一个分块大小
+                try:
+                    response = self.__session.post(f'{url}/{len(chunks_data)}', timeout=30,
+                                       data=chunks_data, headers=headers).json()
+                except requests.ConnectionError:
+                    raise RuntimeError(url)
+                parts.append({"index": i, "ctx": response['ctx']})
+                print(f'kodo: {(i + 1) / chunks:.1%}')  # 输出上传进度
+        self.__session.post(f"{endpoint}/mkfile/{total}/key/{base64.urlsafe_b64encode(key.encode()).decode()}",
+                            data=','.join(map(lambda x: x['ctx'], parts)), headers=headers, timeout=10)
+        r = self.__session.post(f"https:{fetch_url}", headers=fetch_headers, timeout=5).json()
+        if r["OK"] != 1:
+            raise Exception(r)
+        return {"title": splitext(name)[0], "filename": bili_filename, "desc": ""}
+
+    def upos(self, filepath, query):
+        total = os.path.getsize(filepath)
+        with open(filepath, 'rb') as f:
+            name = f.name
+            ret = self.__session.get(
+                f'https://member.bilibili.com/preupload?name={quote(name)}&size={total}&{query}', timeout=5).json()
             chunk_size = ret['chunk_size']
             auth = ret["auth"]
             endpoint = ret["endpoint"]
@@ -365,25 +254,21 @@ class BiliBili:
             # 开始上传
             parts = []  # 分块信息
             chunks = math.ceil(total / chunk_size)  # 获取分块数量
-            for i in range(chunks):  # 单线程分块上传，官方支持三线程
+            for i in range(chunks):
                 chunks_data = f.read(chunk_size)  # 一次读取一个分块大小
-                try:
-                    self.__session.put(f'{url}?partNumber={i + 1}&uploadId={upload_id}&chunk={i}&chunks={chunks}'
-                                       f'&size={len(chunks_data)}&start={i * chunk_size}'
-                                       f'&end={i * chunk_size + len(chunks_data)}&total={total}', timeout=30,
-                                       data=chunks_data, headers={"X-Upos-Auth": auth})
-                except requests.ConnectionError:
-                    raise RuntimeError(url)
+                self.__session.put(f'{url}?partNumber={i + 1}&uploadId={upload_id}&chunk={i}&chunks={chunks}'
+                                   f'&size={len(chunks_data)}&start={i * chunk_size}'
+                                   f'&end={i * chunk_size + len(chunks_data)}&total={total}', timeout=30,
+                                   data=chunks_data, headers={"X-Upos-Auth": auth})
+
                 parts.append({"partNumber": i + 1, "eTag": "etag"})  # 添加分块信息，partNumber从1开始
                 print(f'{(i + 1) / chunks:.1%}')  # 输出上传进度
-
-        prefix = splitext(name)[0]
         r = self.__session.post(
             f'{url}?output=json&name={quote(name)}&profile=ugcupos%2Fbup&uploadId={upload_id}&biz_id={biz_id}',
             json={"parts": parts}, headers={"X-Upos-Auth": auth}, timeout=5).json()
         if r["OK"] != 1:
             raise Exception(r)
-        return {"title": prefix, "filename": splitext(basename(upos_uri))[0], "desc": ""}
+        return {"title": splitext(name)[0], "filename": splitext(basename(upos_uri))[0], "desc": ""}
 
     def submit(self):
         if not self.video.title:
