+++ engine/plugins/upload/slider/__init__.py
